Fix: Split HTTPS Client Tests into Net and Non-Net
The "Net" tests require Internet access (access to google.com etc.);
they are excluded from the tests when executing on Launchpad.
